backend/requirements.txt
# Server packages
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
python-multipart==0.0.6
aiofiles==23.1.0
Pillow==10.0.1

# Client packages
PySide6==6.8.2
asyncio==3.4.3
qrcode==7.4.2
requests==2.31.0

# Common packages
python-dotenv==1.0.0
uuid==1.30

# Testing packages
pytest==7.4.0
httpx==0.24.1

fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
websockets==11.0.3
python-dotenv==1.0.0
pydantic==1.10.7